+++ js/constants.js
@@ -1,0 +1,78 @@
+const LIKES_MIN = 15;
+const LIKES_MAX = 200;
+const COMMENTS_MIN = 0;
+const COMMENTS_MAX = 30;
+const AVATAR_MIN_INDEX = 1;
+const AVATAR_MAX_INDEX = 6;
+const NUMBER_OF_PHOTOS = 25;
+const PHOTO_INDEX_MIN = 1;
+const PHOTO_INDEX_MAX = 1000;
+const DESCRIPTION = [
+  'фото парка с прудом',
+  'фото указателя к морю',
+  'фото морского берега',
+  'фото девушки с фотоаппаратом в руке на пляже',
+  'фото супа с рисовым человеком',
+  'фото чёрной машины с вертикально открывающейся дверью',
+  'фото блюда с клубникой разрезанной на две части',
+  'фото двух стаканов с напитками',
+  'фото самолёта пролетающего над пляжем',
+  'фото выдвигающейся полки для обуви',
+  'фото проход к морю',
+  'фото белой ауди',
+  'фото ресторанного блюда',
+  'фото кота в виде начинки суши',
+  'фото домашних тапочек в виде сапог',
+  'фото самолёта пролетающего над горами',
+  'фото хора',
+  'фото коллекционной машины',
+  'фото тапочек с фонариками',
+  'фото двор с пальмами',
+  'фото азиатского блюда с лаймом',
+  'фото моря на восходе',
+  'фото краба',
+  'фото феерверков на концерте',
+  'фото внедорожника рядом с бегемотом в луже'
+];
+
+const AUTORSNAME = [
+  'Артём',
+  'Александр',
+  'Екатерина',
+  'Дмитрий',
+  'Анна',
+  'Сергей',
+  'Мария',
+  'Алексей',
+  'Ольга',
+  'Михаил',
+  'Наталья',
+  'Андрей',
+  'Татьяна',
+  'Игорь',
+  'Юлия',
+  'Владислав'
+];
+
+const MESSAGES = ['Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+export{
+  LIKES_MIN,
+  LIKES_MAX,
+  COMMENTS_MIN,
+  COMMENTS_MAX,
+  AVATAR_MIN_INDEX,
+  AVATAR_MAX_INDEX,
+  NUMBER_OF_PHOTOS,
+  DESCRIPTION,
+  AUTORSNAME,
+  MESSAGES,
+  PHOTO_INDEX_MIN,
+  PHOTO_INDEX_MAX
+};
